+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `HasError` property to `CliCommandResult`.
+- Add `WithWorkingDirectory(string)` method to `ProgramCli` and `ProgramCli<TCli>`.
 
 ## [1.1.0] - 2021-06-24